Fixed incremental update test, added request retry
Added a clock as a new constructor argument for the OsvDownloadTask to allow testing the incremental update.
The clock is used in the Instant.now() call which allows passing a fixed clock in the test case. Normally the default systemUTC clock is used.
Additionally added another retry config to allow retrying http request for certain response status codes.

Signed-off-by: jonbally <19593213+jonbally@users.noreply.github.com>